Handle windows line endings when running on linux.
Running this on linux with a windows line ending file will fail the
regex, as $ is os dependent. Until now the script has been running on
Windows with Windows line endings, or on Linux (running rspec local)
with linux line endings. Trying to get it running in docker on linux and
this was failing to find any includes.